requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
lnetatmo==4.2.0
    # via sensorthings-utils (pyproject.toml)
pydantic==2.10.6
    # via sensorthings-utils (pyproject.toml)
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.1
    # via sensorthings-utils (pyproject.toml)
pyyaml==6.0.2
    # via sensorthings-utils (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
